subconscious_ai.py/requirements.txt
numpy
colorama
pillow
pygame
librosa
opencv-python
chromadb
sentence-transformers
pyaudio
llama-cpp-python
torch
torchvision
